+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 - NEW: Add "widont" text transformations with
   [typogr](https://github.com/ekalinin/typogr.js).
 - BUGFIX: Ensure consistent ordering of `extraDocs`.
+- BREAKING: `@icons` annotation only requires path to a folder of SVG icons --
+  [#47](https://github.com/oddbird/sassdoc-theme-herman/issues/47)
+- BREAKING: Replace `minifiedIcons` option with `customHTML` option
+  (only included in rendered `@example` annotations).
 
 
 ## 1.0.0-alpha.15: 2017-10-30